[Commands] Remove /wild
Getting this command to work properly to find a safe teleport target
would require a bunch of chunk loading; we’ve tried that in the past,
and it tanked performance *way* too much; since then, the command has
been basically useless because it would generate positions we couldn’t
teleport to because the chunks weren’t loaded.

It also is doubly useless now that a large portion of the map has been
claimed, and I’d rather people explore anyway. Also, the possibility of
someone randomly finding something they’re not supposed to find and that
would have been very hard to find otherwise is something I never liked
about this command anyway.

In sum, this command is stupid and broken, yeet it; good riddance.